undo: Move undo execution under data.undo
Command checkpoint handling and redo execution now live in the
data.undo.checkpoints module inside the package. Undo execution and the
undo command still use the top-level data.undo_checkpoints module.

The remaining module exposes a second location for checkpoint operations
and prevents data.undo from providing the complete public undo and redo
API.

This commit moves undo execution beside redo, updates the undo command, and
removes the top-level data.undo_checkpoints module.

The final undo-package commit will move command and packaging tests to the
new modules and reject the removed top-level imports.
